Real-or-Fake-Image-Classifier/api/requirement.txt
fastapi
torch
torchvision
uvicorn
pillow
